requirements.txt
crewai
crewai_tools
load_dotenv
langchain-groq
google-generativeai
assemblyai
youtube-transcript-api
yt_dlp
whisper

#
langchain-google-vertexai
gpt4all
pyautogen
tqdm
ipywidgets
PyPDF2
streamlit

watchdog
langchain-chroma
flashrank